Complete refactor.
Added new rest api client.

+++ Wpf/Gracjan/CarRentalWPF/CarRentalWPF/Models/CarModel.cs
@@ -30,7 +30,7 @@ namespace CarRentalWPF.Models
 
         public string VIN { get; set; }
 
-        public int PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
 
         public string FullName
         {
